added strafe and straight commands

+++ src/org/usfirst/frc/team4611/robot/subsystems/DriveTrain.java
@@ -19,14 +19,14 @@ public class DriveTrain extends Subsystem {
 		RobotMap.driveTrain.drivePolar(mag, angle, z);
 	}
 	
-	public void motionMagicStraight(int positionUnits) {
+	public void motionMagicStraight(double positionUnits) {
 		RobotMap.driveTrainBL_Talon.set(ControlMode.MotionMagic, positionUnits);
 		RobotMap.driveTrainBR_Talon.set(ControlMode.MotionMagic, -positionUnits);
 		RobotMap.driveTrainFL_Talon.set(ControlMode.MotionMagic, positionUnits);
 		RobotMap.driveTrainFR_Talon.set(ControlMode.MotionMagic, -positionUnits);
 	}
 	
-	public void motionMagicStrafe(int positionUnits) {
+	public void motionMagicStrafe(double positionUnits) {
 		RobotMap.driveTrainBL_Talon.set(ControlMode.MotionMagic, -positionUnits);
 		RobotMap.driveTrainBR_Talon.set(ControlMode.MotionMagic, -positionUnits);
 		RobotMap.driveTrainFL_Talon.set(ControlMode.MotionMagic, positionUnits);
